Change `makeNewBlock` interface to accept the pre-images
As a pre-requisite to store the pre-images in the block,
we need to change `makeNewBlock`'s interface to accept said pre-images.
This also simplifies the API quite a bit, by replacing 3 parameters
with a single one.

In the process of doing so, and in order to make the tests more future-proof,
the 'unit' integration tests now use "real" pre-images, which required to use
`agora.utils.Test`, which links in `sqlite3` indirectly, hence the change
to the build script.