Fix #2284: Schnorr signatures misusage leaks private keys
This effectively changes the formula used for signing blocks, turning it upside-down:
Instead of offsetting the 'R' by the pre-image as a point,
we commit to the pre-image being the 's' of the signature,
and provide 'R' such as 'S == c * R + X'.
In order to find the adequate 'R' value, one has to know both 'x' and 'r',
as one may only derive CR from the known values, and CR cannot be divided by 'c'.
In the process of this change, the no-longer-used 'getCommitmentNonce' was removed.