Tests: Replace `spendable` with `getUTXOs`
To ensure that tests are more reliable and not effected by empty blocks.
When getting spendable outputs a node can be requested to provide the
utxos with sufficient funds for the intended transactions in the tests.
Also making use of `generateBlocks` and `postAndEnsureTxInPool` to make
the test code more readable and robust.