Trivial: rename test function `getUTXOs` to `getSpendables`
This function is intended to return unspent utxos for using in unit test
code. It will be updated in the next commit to check that it really is
spendable in the next block. Note that the unit test that includes it
was previously disabled so a future commit will fix it and enable it
again.