Tests: Add `freezeUTXO` to base test class
This will generate frozen utxo by using the new `getSpendables` on the
first client to fetch unspent utxo from the utxo_set of that client. The
frozen utxo are output to each of the nodes indicated by indexes
provided so that they will be able to enroll in a future block.