Tests: Ensure required signature count in `BlockRewards`
In the last test in this file we could not use quorum_threshold of 100%
which is how the other tests ensure that all signatures are present. For
this test the signature count is different for even block heights. To
prevent ocasional failures the test now waits after each block for the
desired signature count so that the calculations are correct for the
expected rewards and fees.