Manager/RPC: Allow new RPC connections to be merged to existing ones
When a remote initiates a new RPC connection, we will try to discover them
as an Agora node. In the case where the remote is actually an Agora node, NetworkManager
will try to merge that connection into an existing RPC connection. This way incoming connections
can be used for outgoing requests as well.