Improve README: Update test command and build instructions
Use the config file in devel to get a full, working example.
Move the GC section out of the README as it is intended for developers and advanced usage.
Re-state in the build instructions that the dependencies need to be acquired before,
as one might just copy-paste those and expect it to work.
Add minimum required version of libsodium.