Ledger: Revert a test that is no longer useful
This was aimed at being a simple test for a valid block.
However, the Ledger code has evolved so much that generating a valid block
is no simple task nowadays, and multiple helper functions are exposed.
A proper test would need to do everything "manually",
which will not bring much value, as it is already tested by network unittests.