Trivial: Use non-deprecated syntax for delegate literal
In D, delegate literals are either '() {}' or '() => expression',
but other languages use '() => {}', which makes it very easy to get wrong.
As a result, using '() => {}' now triggers a deprecation message to catch such error.
In our case, we were using '() => {}()' so the usage was actually correct,
but unnecessarily obscure. Removing the '=>' and trailing parentheses have the same effect.