Fix #2624: Implement support for DNS name compression
The RFC specifies that conformant implementations may not generate
compressed names, but must be able to understand them.
As we are moving towards having the DNS code be a consumer of messages,
and not only a producers, support is becoming a blocker.

As a result, we loose the ability to deserialize individual structs,
and deserialization must only happen at the `Message` level.