DNS: Make Header.ID unsigned (short)
It was originally signed because the RFC did not mark it as unsigned
(it did mark other fields as explicitly unsigned), however it is an
identifier (not a counter), hence having it either signed or unsigned
should not matter: Except it does.
When comparing it against a known value, or a different type,
comparison will fail for '==' if the value is negative.
Using 'is' for the comparison would behave correctly,
but only for types of the same size.
To avoid those potential pitfalls, using ushort is the sanest approach.