Fix #2320: Improve address registration UX
This removes 'addresses_to_register' from validator, and instead add it to
the related entry in 'interfaces'.
It simplifies things for the user as the port does not need to be repeated,
nor the scheme, and when adding / removing interfaces, only one place
needs to be edited. The port was also made optional now that we have default ports.