DNS: Add a factory function for `ResourceRecord`
`ResourceRecord`s are the core of the DNS system, and are essentially
an `enum` type when framed from a D point of view.
However, they are currently defined following the wire format,
which is less than an ideal API for the user.
As a step towards having a better API for them, a factory function needs
to be introduced, hiding the underlying details / boilerplate.