Cache the Coinbase tx
During nomination the Coinbase transaction is for the payout period
preceding the current payout period. This means that no more signatures
are added to these blocks as the catchup only signs those within the
current payout period. Therefore we only need to calculate the rewards
and fees to be paid a single time and then use the cached result.