Validator: Reveal preimages on new block events
On enrollment blocks, network is missing preimage of the newly
enrolled validator for the next block. In our production environment,
our block time is short enough that sometimes reveal timer is not triggered
before creation of the next block, resulting in the newly enrolled validator
to be slashed immediately.

onPreImageRevealTimer seems like re-entrant so this should not be a problem
that we call it from outside of the timer.