D-Scanner: Disable 'could be immutable' check
This check currently triggers 100/132 times.

This check is not very useful to us (it differs greatly from our coding style),
but is also quite buggy: For example, as D-Scanner doesn't do semantic analysis,
it ends up triggering in places where the variable cannot be made const/immutable,
because it is then used in a context where it needs to be mutable (e.g. struct initializer).
See D-Scanner issue 847 for an example.