Pre-image: Correct test-only isValid function and rename parameters
The function was extremely misleading: It was taking the images as parameter,
and passing it to 'isInvalidReason' with the parameters swapped.
However, since the call site also swapped parameter (e.g. 'new_preimage.isValid'),
this worked. Correcting the parameter names ensure that future users are not thrown off.

Additionally, parameters were renamed to avoid including the type name.