Implement toHash whenever opEquals is implemented
Because AA use both toHash and opEquals, implementing only opEquals
runs the risk of creating bugs with associative arrays.

An example would be a struct which has two members, but only compares one for equality:
```
struct Foo
{
    int a; // Only 'a' is used in `opEquals`
    int b;
}
```

In that context, if `opEquals` is implemented but not `toHash`,
the compiler will generate a `toHash` automatically, which will most likely
put two equal structs (by `opEquals` definition) into separate bucket,
creating duplicates.

Later on, when the AA is looked up for a specific value, two things can happen:
- The `Foo` instance looked up uses a value for `b` which results in a hash that
  doesn't end up in the same bucket as one of the two previously added values,
  hence the lookup will fail (the AA will believe it has no such value);
- The `Foo` instance looked up uses a `b` value that makes it fall within one of
  the bucket where an instance exists, and this instance will be returned.

Which instance will be returned is arbitrary and thus, for AA to work properly,
we must always make sure `toHash` is implemented.