Signature: Introduce OmitSingle
This will let us sign multiple outputs while letting one of them
to be open to modifications.

Need for this comes from the settlement TXs in Flash. Settlement
may include multiple outputs for the channel funds which are signed
at the time of the micro-transaction and we may need to add a refund
output for the inputs we add to fund the fees for the externalization
of the tx.