NetworkManager: Merge two related arrays into a struct
We were doing some dependent bookkeeping with the apis and addresses,
and were assuming that two items at the same index were related.
This is a fairly obvious sign that the arrays must be merged.
In the process, the useless `apis` arguments passed to attemptRequest
was dropped, as it was always the same, was already available via the class,
and the same link with addresses was assumed anyways.