Bump required LDC to v1.28.0
As we are living on the bleeding edge, we frequently end up relying on
brand new features, and/or bugfixes/features we've ourselves upstreamed.
To that extend, it would be quite a bit of work, and not so useful,
to maintain a large backward compatibility.

As a result, `ldc-master` had to be disabled, as it's currently ldc-1.28.0+,
and there's a dub bug which makes it so it doesn't recognize it correctly,
leading to an error about unsatisfied toolchain dependency.