Allow setting quorums while nominating
On a block being accepted, a current nomination round stops, and
the newly constructed quorums are set. But there might be a situation
where the previously called nomination process is running while
setting new quorums, which is a probable situation but is not accounted
for in our code. So we allow setting new quorums even while a previous
called nomination is running if the nomination is for the next height.