Implement Chain ID
Chain ID is hashed together with all user data which makes
signatures for different networks to be unique.

Initial idea was to include genesis hash instead of a magic number
but that creates a chicken&egg problem since genesis hash of different
network with exact same genesis block should be different as well.

Fixes #2728