Refactor the handling when missing tx from other nodes
If we recieve an envelope with tx set containing any tx we do not yet
have in the pool we were throwing away the envelope.
Now we will still process the envelope but when SCP calls our
`ValidateValue` we return `ValidationLevel.kMaybeValidValue` which is
intended for the purpose of missing context to fully validate.