Enable UnlockNotify for d2sqlite3
If our DB is locked by another process, we should gracefully wait
for it to be unlocked instead of erroring right away.

This will help NODE0 in production, DB of which is also accessed by
Grafana.

Homebrew sqlite has the native support disabled so we use the fake
UnlockNotify provided by the D bindings.