DNS: Prepare for `@disable`d `Domain` ctor
Disabling a `struct` default constructor means that it disables
default construction for any struct that include an instance of this `struct`.
As we are going to disable `Domain`'s default ctor in the next commit,
we prepare this one by explicitly instantiating some `struct` / `union`
that include a `Domain` in them.