Introduce NodeLedger as an intermediate Ledger type
A NodeLedger is a more involved Ledger than the basic Ledger,
as the basic Ledger should be usable by a client without some of
the machinery present in the NodeLedger, most notably the two pools.

As a first step, and instant win to decouple things, we can move
`onValidatedBlock` to the `NodeLedger`, as a client would most likely
perform operations sequentially (call `acceptBlock` then check state).