Ledger: Move EnrollmentManager out of the base Ledger
EnrollmentManager should be only needed for users that want to validate the chain.
However, due to an unfortunate dependency between the Enrollment pool and it,
it's currently instatiated in the Full Node.
Moving it out of the base Ledger class ensures that at least library users
need not to bother with it.