Ledger: Use public interface over internal ones in tests
`lastBlock` is a public wrapper for `last_block` and should be used,
so that we can move the unittest around freely without it breaking.

`acceptBlock` is the entry point to add a block,
and even if we're testing the internals (where something throws),
we can and should use the public entry point to minimize
the risk of accidentaly breakage in the future.
The `assert` also ensures that the test doesn't pass for the wrong
reasons (passing an invalid `Block`).