Quorum unit test: Simplify test code
The code was looping over a single output within a transaction so the
loop can be removed. In fact the transaction is not required as only the
output is used in the test UTXO set. The lookup hash for the storage is
generated only from a passed in test id, the index of the node and the
amount.