Quorum unit test: Fix incorrect order of nodes
The tests that compare the expected output with the actual output were
not using the intended order. When the stakes on the first and last node
are much higher the output of the first and last in the array of counts
should also be much higher. The tests are also refactored to check this.