ValidatorCleanRestart: Re-enable the first test
The test also required some changes as it had been disabled for some
time and was out of sync with the implementation code. The actual change
in the test that prevents it from being unreliable is waiting for the
new set of validators to complete network discovery before generating
the new block.