Explicitly transfer target height of an Enrollment
Previously, target height was assumed to be the next height. After
next block is externalized, enrollment was removed from the pool, even
if not externalized.

If an Enrollment gets gossiped after externalization of it's target
height that the originating node intended, it could be externalized
at a later height. Which caused the commitment in the Enrollment to be
associated with a wrong height and made impossible to reveal any new
PreImages.

By sending target height explicitly along the enrollment, we guarantee
that a delay in externalization of Enrollments don't doom the Validator.

Fixes #2838