Ensure `acceptBlock` is called when block is fetched
Do not call `ledger.acceptBlock` directly but call the function which
also ensures the validators who are expired next block are removed from
the whitelist and the active validators can be set as whitelisted.
This also fixes an out by one error in the logic for determining which
validators have expired.