Ledger: Move FeeManager / UTXOCache instantiation in the Ledger
The instantiation happening in the node is a leftover of the time
where dependency injection was done at the class level.
Now that we just pass a ManagedDatabase, there is no need for it.
If tests want to implement a specific behavior for one of those
classes, they can re-introduce the dependency injection in the Ledger.