No Agora override for `SCPDriver::computeHashNode`
The Agora `override` had been added to provide randomness by using the
hash of the previous block header. However, the `prev_value` param
passed to `nominate` can achieve the same result. With the advantage of
preventing a potential race condition where the ledger height is not one
less than the `slot_idx` when executing `computeHashNode`. This could
occur if a block is fetched by catchup during a nomination round.