Ledger: Do not error/throw when re-adding the HEAD block (esp. Genesis)
When Faucet was switched over to use a Ledger,
it was found that trying to add genesis to the Ledger
results in an Exception being thrown.
While Faucet could be changed not to do that,
it seems like a common (yet harmless) error would be
to attempt 'acceptBlock' with the Ledger HEAD.
This is now handled more gracefully.