Network Unit tests with subset of Genesis validators
This makes it possible to configure the number of validators used in a
test. This can sometimes help to debug a problem or make the test run
faster. The Simple test is updated to run with 4 nodes with quorum size
of 3. Some of the longer running tests can be sped up quite a bit.