RPC: Keep unmerged connection alive
We were closing the connections that we could not
merge due to the maximum connection limit per RPCClient.
Hoping that the peer will reuse the connections we started,
so that we can not have more than configured number of connections
per client. But it is subject to a number of race conditions.

This could affect stability of the communication as it
might result in too many connections to be discarded and
re-established.

I haven't witnessed this causing any problems but I would rather
be safe than sorry.