Flash: Increase retry delay of some tests
The only difference those tests' config had with the default config of flash tests
was the retry_delay, which was 10 msecs instead of 100 msecs.
Due to the variability of the runners (either dev machines or CI machines),
and the small window of times involved, this either makes the test flakey or
has is not part of the test. Hence, we simply use the default config now.