Fix flash unittest crashes on CI
There was some situation where a necessary `update tx` does not reach
validators because some block for the UTXO that an `update tx` refers
to has not yet been externalized before the tx is received. It is
because the unittests in the Flash make blocks without transactions
so a block can be externalized with the necessary txs. So we remove
the cause of missing txs.