Ledger: Dont validate individual TXs when preparing a set
A new TX set is prepared periodically in checkNominate(), and this
is a costly operation. Rely on Nominator to validate the prepared
ConsensusData in it's own prepareNominatingSet which uses memoization.

If we encounter an invalid TX while validating a value, we should also
remove it from the pool so that we won't keep trying to nominate invalid
values.