Make TimePoint a struct instead of an alias
Using a single integer type for time point causes some problems,
and instead of sticking with a bad C habit, we should use a full
typed approach like std.datetime does.
In order to do so, the first step is to remove assumptions that a ulong
converts to a TimePoint. The next step will be to remove the assumption
that a TimePoint converts to a ulong.