NetMan: Delay NetworkClient instantiation to remove tryMerge
By delaying the NetworkClient instantiation, we can actually simplify the code,
removing the need for the tryMerge function and making the code easier to follow.
This also paves the way to remove the duplication between clients and required_peers,
as in the future we would like to have connection-less NetworkClient that handles
their own 'ConnectionTask'.