Trivial: Unify dependency injection functions under the 'make' prefix
Originally, to do dependency injection, we were using the 'get' prefix,
which was a mistake in hindsight. It has been changed in a few places,
but some usages remained, which has now been cleaned.
Note that in the network unittest, we still (consistently) use the 'create' prefix
to note initial node instantiation (which is not subject to dependency injection).