Signature Catchup for payout period
The validator reward payouts are reduced for blocks where the signature
count is less than 100%. This means that if nodes have a different
number of signatures for those blocks then the Coinbase payouts will be
different. To recover from this we can trigger signature catchup when
this happens after resetting the cached Coinbase tx.