Temporarily set node.testing to true by default
As we are preparing the TestNet release, we still have a few usability
improvements needed for users to be able to run a node smoothly.
Anything in the configuration file will be off-putting, especially with Docker,
as a volume / mount is needed. Additionally, the CoinNet Genesis block is broken
anyways, and needs to be re-done, so for the time being, TestNet is the default.