Registry: Remove table drop in case of type change
This piece of code was dropping the registry tables iff the type of the registry
was changed between two runs. There's apparently no technical reason to do so,
as all types use the same table format.
This code created problems because it consistently prints 3 error messages when a node
starts, which can be startling to users. This could be avoided by checking for the
table existence before actually doing the SELECT, however, when considering the logic,
it stood out as no other node has that behavior (e.g. switching between full node and validator).
In this case, I think it is better to simply not do this and let the operator
wipe the table clean (by removing cacheDB) if that behavior is desired.