Fix #2989 - Revert PR #2981
PR #2981 assumed that `registry_address` contained the address of a DNS server,
irrespective of the actual schema. This broke the live system as DNS request
are now sent over port 80. Revert to the default DNS resolvers,
as it will work properly, although it will issue wrong requests
in integration tests.