Flash test: no block at requested height
A flash test sometimes fails during `gossipChannelsOpen` if the call to
`getBlock` fails. This can happen if the discovered channel height is
greater than the current block height of the peer we attempt to fetch
from. By storing the last block rather than block height we can help
prevent this sort of race condition.