Improve TestNet configuration and allow dynamic ConsensusParams
This fixes two major issues with the current TestNet setting:
- A network reset requires a new release and users to manually upgrade;
- Using '--testnet' does not work when a one wants to run a validator;

We fix both issues by first adding a new endpoint to the API, which returns the ConsensusParams.
The current approach is not ideal as ConsensusParams and ConsensusConfig are slightly different,
but it will cover our immediate need. The second issue is fixed by only overriding the consensus
section whenever we detect there is a configuration file we can read.
The realm, the registry address, and node.testing still need to be set correctly,
but the sample config file does this and people are likely to start off of it.