Clock: Remove validator special-casing
The Clock was handled a bit bizzarely: The interface was designed for
a validator, but then disabled in tests and FullNode.
This is a red flag that the interface is not suitable.
In order to remove the amount of special casing, we're moving the
clock synchronization to FullNode so that 'start' is always called,
even in unittests.