Validator: Reveal preimages according to expected height
When network falls behind on blocks, it creates a lot of blocks
in a short span and this results in some nodes to be slashed in the
process. If we reveal the preimage of the expected height instead
of the ledger height, we should decrease the chance of us getting
slashed in this scenario.