Bump serialization and fix memory corruption in URI deserialization
The previous code was retaining a reference to the stack,
because it did not dup the result like the serializer would.
It also failed to properly validate the string before passing
it to Address's constructor, which could trigger assert or
other logic errors. serialization was changed to introduce
deserializeArray as a simple way to handle this case.