`ValidatorCleanRetsart` test: remove sleeps
The first test was relying on a sleep before generating the last block.
This was to give time for discovery of the new set of validators, so
that in the `addBlock` function the check for the tx gossip to each
client of the new active validators would pass.
However in reality if the tx gossip failed it should be fetched during
nomination as it will be discovered as missing in the nominated tx set.
The test is now updated to test that this actually works as expected.