`checkExternalize`: return early if already externalized
It can be that several signatures are received at the same time and the
`updateMultiSignature` calls the `checkExternalize` multiple times to
check if we have a majority. If one of the checks passes then the
`pending_block` is reset to `Block.init` and subsequent calls to
`checkExternalize` should then be ignored.