Test.base.d correction to waitForPreimages logging
Previously the index and node address at that index of the enrollment
was logged which may not be the actual index of the node.
For example, the first node does not enroll for some reason (perhaps
the test case makes the node enroll with invalid data). The first index
of the enrollments in the header is now related to the node at index 1
not 0. To be always accurate the utxo is logged.