QuorumPreimage unit test: Do not use GenerateBlocks
For blocks 21 -> 39 it is better to send the tx to one client and check
for the height in the test code as there are outsider validators. In the
`generateBlocks` function it checks that the txs are gossiped to all the
validators before setting the time for the next block. But as the
outsiders were not enrolled until block 20 the client may gossip
the tx before it has updated the peers it sends to. Normally this does
not matter because once an envelope is recieved by the newly enrolled
node it will try to fetch it.