Coinbase: reset cache if signature is added
It can happen that the nodes do not all have the same signatures in the
blocks for the payout period. In the case where a signature is added to
the block header of a block within the payout period we should reset the
cached version of the Coinbase so that it can be re-calculated later.