Remove Ocean dependency, use dtext instead
We were mostly using Ocean for its Formatter / Logger facily.
But Ocean is a lot of code and contains a few module constructors
that could bite us. Removing our dependency on it in favor of dtext
ensures that the odd update process we have to go through
(as Ocean is Linux-only) is simplified.