UnlockAge test: use 2 test validators
No need to test with more than two as we are verifying the transaction
input unlock not consensus. It could even be just one but that would
require a bit more test refactorring as currently it fails to set up
test network in unit tests if only a single validator is used.