Add new validator to whitelist
In the case of the genesis validators, there is no address registered right
after nodes start from scratch, which means the genesis validators might be
banned although there have been the `whitelist` calls for the validators in
the middle of accepting the Genesis block. So we should add the UTXOs to
the `whitelist` when the handshaking has been completed.