Validator: Move addresses to register to a common node configuration
It was required to provide several URL addresses to be registered as
public interface for a Validator node. This configuration method was prone
to invalid configurations (i.e. schema and port mismatches) and it was
incapable of registering correct schema and port for default (public IP)
configuration. Address registry will be handled through public interfaces
according to the configuration.