Resend latest envelopes if half block late
Rather than waiting a full block interval to start resending envelopes
with same nomination it can help prevent delayed blocks if we only wait
for half the duration. This will probably need to be optimized at some
point but first we want to ensure we do not get blocked when peers are
not connected when we first sent the envelopes.