EnrollmentManager: Merge setCommitment and setEnrollmentKey, remove test
The unittest was testing a simple setter/getter, and the likelihood of this
ever getting broken was very low.
It was also putting the `EnrollmentManager` in an inconsistent state.
In order to prevent this inconsistent state, both `set` methods were merged,
and the test was removed for its lack of actual usefulness.