Resend envelopes at exactly half block time
Change to greater than or equals to expected block time plus half block
interval. This is only to enable the unit tests to benefit from the
resend as the time set is exact and does not move forwards. Although it
would be nice to implement the tests to start at expected time and
progress to intervals within the block time span. That is for another
day.