Only add signature if not slashed
If a block is being externalized that does not contain a preimage for a
validator we must not add a signature for that validator. The validator
is being slashed at this point so it is too late to add the preimage and
signature, otherwise the block hash would change and the combined
signature would become invalid.