Engine: Fix hasScriptFailed() behaviour
docs say;

The script is considered sucessfully executed only if its stack
contains exactly one item, and that item being `TrueValue`.

Which was what I thought the implementation did too. But the
implementation only checked if stack had a TrueValue on the top
of the stack, ignoring if it had more data.